Clean: remove doctrees folder
Per https://stackoverflow.com/questions/33904042/is-doctrees-folder-required-for-displaying-html-docs-with-sphinx

doctrees contains a cache of rst source docs and are not needed in the
repository.

These files are getting large enough to trigger warnings during git
push:

```
File docs/doctrees/mooring_example. doctree is 58.16 MB; this is larger than GitHub's recommended maximum file size of 50.00 MB
```

To mitigate this warning and further errors we are ignoring the
`docs/doctree` and the older `docs/.doctree` folders.